Add transparency fixing option
Add a processing option to fix any area that should be transparent but was incorrectly rendered white.

Modularize the code so that each processing option has its own class.

Remove the dependency on Guava and instead using code copied from Android to parse colors.

Add a dependency on commons-cli and use it to handle all command line arguments processing.